WIP: add code + array format
restructured parts of the code for better overview
