+++ lib/index.js
@@ -1,1 +1,6 @@
-module.exports = require('./key-handler');
+/* @flow */
+import KeyHandler from './key-handler';
+export { KEYDOWN, KEYPRESS, KEYUP } from './constants';
+export { default as keyHandleDecorator, keyToggleHandler } from './key-handle-decorator';
+
+export default KeyHandler;
