+++ lib/key-handler.js
@@ -5,19 +5,31 @@ import PropTypes from 'prop-types';
 import {canUseDOM} from 'exenv';
 
 import {KEYDOWN, KEYPRESS, KEYUP} from './constants';
-import {isInput, matchesKeyboardEvent, eventKey} from './utils';
+import {type rawOrArray} from './types';
+import {isInput, matchesKeyboardEvent} from './utils';
 
 type Props = {|
-  +keyValue?: string,
-  +keyCode?: number,
+  +keyValue?: ?rawOrArray<string>,
+  +keyCode?: ?rawOrArray<number>,
+  +code?: ?rawOrArray<string>,
   +keyEventName: KEYDOWN | KEYPRESS | KEYUP,
   +onKeyHandle?: (event: KeyboardEvent) => void,
 |}
 
 export default class KeyHandler extends React.Component<Props> {
   static propTypes = {
-    keyValue: PropTypes.string,
-    keyCode: PropTypes.number,
+    keyValue: PropTypes.oneOfType([
+      PropTypes.string,
+      PropTypes.arrayOf(PropTypes.string),
+    ]),
+    keyCode: PropTypes.oneOfType([
+      PropTypes.number,
+      PropTypes.arrayOf(PropTypes.number),
+    ]),
+    code: PropTypes.oneOfType([
+      PropTypes.string,
+      PropTypes.arrayOf(PropTypes.string),
+    ]),
     keyEventName: PropTypes.oneOf([KEYDOWN, KEYPRESS, KEYUP]),
     onKeyHandle: PropTypes.func,
   };
@@ -30,12 +42,15 @@ export default class KeyHandler extends React.Component<Props> {
     return false;
   }
 
-  constructor(props) {
+  constructor(props: Props) {
     super(props);
 
     /* eslint-disable no-console */
+    if (props.keyCode) {
+      console.warn('The `keyCode` is deprecated, use `code` instead');
+    }
 
-    if (!props.keyValue && !props.keyCode) {
+    if (!props.keyValue && !props.keyCode && !props.code) {
       console.error('Warning: Failed propType: Missing prop `keyValue` or `keyCode` for `KeyHandler`.');
     }
 
@@ -59,7 +74,8 @@ export default class KeyHandler extends React.Component<Props> {
   }
 
   handleKey = (event: KeyboardEvent): void => {
-    const {keyValue, keyCode, onKeyHandle} = this.props;
+    console.log('...asd')
+    const {keyValue, keyCode, code, onKeyHandle} = this.props;
 
     if (!onKeyHandle) {
       return;
@@ -71,67 +87,10 @@ export default class KeyHandler extends React.Component<Props> {
       return;
     }
 
-    if (!matchesKeyboardEvent(event, {keyValue, keyCode})) {
+    if (!matchesKeyboardEvent(event, {keyValue, keyCode, code})) {
       return;
     }
 
     onKeyHandle(event);
   };
 }
-
-/**
- * Types.
- */
-
-type DecoratorProps = {
-  keyValue?: string,
-  keyCode?: number,
-  keyEventName?: string,
-}
-
-type KeyhandleDecoratorState = {
-  keyValue: ?string,
-  keyCode: ?number,
-};
-
-function keyHandleDecorator<T>(matcher?: typeof matchesKeyboardEvent): Function {
-  return (props?: DecoratorProps & T): Function => {
-    const { keyValue, keyCode, keyEventName, ...other } = props || {};
-
-    return (Component) => (
-      class KeyHandleDecorator extends React.Component<T, KeyhandleDecoratorState> {
-        state: KeyhandleDecoratorState = {
-          keyCode: null,
-          keyValue: null,
-        }
-
-        render() {
-          return (
-            <div>
-              <KeyHandler keyValue={keyValue} keyCode={keyCode} keyEventName={keyEventName} onKeyHandle={this.handleKey} />
-              <Component {...other} {...this.state} />
-            </div>
-          );
-        }
-
-        handleKey = (event: KeyboardEvent): void => {
-          if (matcher && matcher(event, this.state)) {
-            this.setState({ keyValue: null, keyCode: null });
-            return;
-          }
-
-          this.setState({ keyValue: eventKey(event), keyCode: event.keyCode });
-        };
-      }
-    );
-  };
-}
-
-export const keyHandler = keyHandleDecorator();
-export const keyToggleHandler = keyHandleDecorator(matchesKeyboardEvent);
-
-/**
- * Constants
- */
-
-export * from './constants';
