+++ lib/types.js
@@ -1,0 +1,2 @@
+/* @flow */
+export type rawOrArray<T> = T | T[];
