+++ lib/utils.js
@@ -1,77 +1,12 @@
 /* @flow */
+import {KEYDOWN, KEYPRESS, KEYUP, KEY_CODE_KEYS, NORMALIZED_KEYS} from './constants';
+import {type rawOrArray} from './types';
 
-import {KEYDOWN, KEYPRESS, KEYUP} from './constants';
-
-/**
- * Types.
- */
-
-type KeyboardKey = {
-  keyValue: ?string,
-  keyCode: ?number,
-};
-
-/**
- * Constants.
- */
-
-const NORMALIZED_KEYS: {
-  [key: string]: string,
-} = {
-  'Esc': 'Escape',
-  'Spacebar': ' ',
-  'Left': 'ArrowLeft',
-  'Up': 'ArrowUp',
-  'Right': 'ArrowRight',
-  'Down': 'ArrowDown',
-  'Del': 'Delete',
-  'Win': 'OS',
-  'Menu': 'ContextMenu',
-  'Apps': 'ContextMenu',
-  'Scroll': 'ScrollLock',
-  'MozPrintableKey': 'Unidentified',
-};
-
-const KEY_CODE_KEYS: {
-  [string: string]: string,
-} = {
-  '8': 'Backspace',
-  '9': 'Tab',
-  '12': 'Clear',
-  '13': 'Enter',
-  '16': 'Shift',
-  '17': 'Control',
-  '18': 'Alt',
-  '19': 'Pause',
-  '20': 'CapsLock',
-  '27': 'Escape',
-  '32': ' ',
-  '33': 'PageUp',
-  '34': 'PageDown',
-  '35': 'End',
-  '36': 'Home',
-  '37': 'ArrowLeft',
-  '38': 'ArrowUp',
-  '39': 'ArrowRight',
-  '40': 'ArrowDown',
-  '45': 'Insert',
-  '46': 'Delete',
-  '112': 'F1',
-  '113': 'F2',
-  '114': 'F3',
-  '115': 'F4',
-  '116': 'F5',
-  '117': 'F6',
-  '118': 'F7',
-  '119': 'F8',
-  '120': 'F9',
-  '121': 'F10',
-  '122': 'F11',
-  '123': 'F12',
-  '144': 'NumLock',
-  '145': 'ScrollLock',
-  '224': 'Meta',
-};
+type KeyboardKey = {|
+  +keyValue: ?rawOrArray<string>,
+  +keyCode: ?rawOrArray<number>,
+  +code: ?rawOrArray<string>,
+|};
 
 /**
  * Check if `given` element is an input / textarea form element or acts as one.
@@ -100,15 +35,34 @@ function isContentEditable(element: HTMLElement): boolean {
  * Matches an event against a given keyboard key.
  */
 
-export function matchesKeyboardEvent(event: KeyboardEvent, {keyCode, keyValue}: KeyboardKey): boolean {
+function elmntOrArrayMatch<T>(a: T | T[], b: T) {
+  if (Array.isArray(a)) {
+    return a.includes(b);
+  }
+  return a === b;
+}
+
+export function matchesKeyboardEvent(event: KeyboardEvent, {keyCode, keyValue, code}: KeyboardKey): boolean {
+  console.log('asdasd', keyValue, keyCode);
   if (!isNullOrUndefined(keyValue)) {
-    return keyValue === eventKey(event);
+    const value = eventKey(event);
+    if (elmntOrArrayMatch(keyValue, value)) {
+      return true;
+    }
+  }
+
+  if (!isNullOrUndefined(code)) {
+    if (elmntOrArrayMatch(code, event.code)) {
+      return true;
+    }
   }
 
   if (!isNullOrUndefined(keyCode)) {
-    // Firefox handles keyCode through which 
-    const code = event.keyCode || event.which;
-    return keyCode === code;
+    // Firefox handles keyCode through which
+    const keyCodeTarget = event.keyCode || event.which;
+    if (elmntOrArrayMatch(keyCode, keyCodeTarget)) {
+      return true;
+    }
   }
 
   return false;
